+++ src/main/java/ru/peytob/mineville/machine/nodes/INode.java
@@ -1,6 +1,5 @@
 package ru.peytob.mineville.machine.nodes;
 
-import ru.peytob.mineville.machine.nodes.Node.ChildException;
 import ru.peytob.mineville.machine.nodes.Node.NodeState;
 
 /**