+++ src/main/java/ru/peytob/mineville/machine/nodes/decorators/LimitNode.java
@@ -2,6 +2,7 @@ package ru.peytob.mineville.machine.nodes.decorators;
 
 
 import ru.peytob.mineville.machine.BehaviorTree.Context;
+import ru.peytob.mineville.machine.nodes.ChildException;
 
 /**
  * Limit node class.