+++ src/main/java/ru/peytob/mineville/machine/nodes/decorators/UntilSuccessNode.java
@@ -1,28 +1,23 @@
 package ru.peytob.mineville.machine.nodes.decorators;
 
-
-import ru.peytob.mineville.machine.BehaviorTree.Context;
+import ru.peytob.mineville.machine.BehaviorTree;
 import ru.peytob.mineville.machine.nodes.ChildException;
+import ru.peytob.mineville.machine.nodes.Node;
 
 /**
- * Repeat node class.
- * Makes its child perform until interrupted.
+ * The node that makes its child perform its tasks until its state is SUCCESS.
  */
-public class RepeatNode extends DecoratorNode {
+public class UntilSuccessNode extends Node {
 
     /**
      * Constructor that sets the link on the context.
+     *
      * @param context context of the tree the node belong to
      */
-    public RepeatNode(Context context) {
+    public UntilSuccessNode(BehaviorTree.Context context) {
         super(context);
     }
 
-    /**
-     * Sends a signal to make child perform.
-     * @return RUNNING
-     * @throws ChildException when decorator has no child
-     */
     @Override
     public NodeState tick() throws ChildException {
         if (children.size() == 0) {
@@ -32,7 +27,7 @@ public class RepeatNode extends DecoratorNode {
 
         NodeState childState = children.get(0).tick();
 
-        if (childState == NodeState.SUCCESS) {
+        if (childState == NodeState.FAIL) {
             children.get(0).setReady();
             state = NodeState.RUNNING;
         } else {
@@ -41,5 +36,4 @@ public class RepeatNode extends DecoratorNode {
 
         return state;
     }
-
 }