+++ src/main/java/ru/peytob/mineville/mas/Ontology.java
@@ -1,0 +1,5 @@
+package ru.peytob.mineville.mas;
+
+/** Ontology class represents the knowledge base of agents. */
+public class Ontology {
+}