+++ src/test/java/ru/peytob/mineville/machine/nodes/decorators/InverterNodeTest.java
@@ -3,6 +3,7 @@ package ru.peytob.mineville.machine.nodes.decorators;
 import org.junit.Assert;
 import org.junit.Test;
 import ru.peytob.mineville.machine.BehaviorTree;
+import ru.peytob.mineville.machine.nodes.ChildException;
 import ru.peytob.mineville.machine.nodes.LeafNode;
 import ru.peytob.mineville.machine.nodes.Node;
 
@@ -35,7 +36,7 @@ public class InverterNodeTest {
             try {
                 root.addChild(new IsDoorOpenNode(context));
             }
-            catch(Node.ChildException ex)
+            catch(ChildException ex)
             {
 
             }